reorganize dto, converter packages
