+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/converter/candidate/CandidateAvailabilityDtoToCandidateAvailability.java
@@ -1,6 +1,6 @@
-package com.github.TheDreigon.JavaInterviewCalendarAPI.dto.converter;
+package com.github.TheDreigon.JavaInterviewCalendarAPI.converter.candidate;
 
-import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.CandidateAvailabilityDto;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.candidate.CandidateAvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.CandidateAvailability;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
