+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/converter/interviewer/InterviewerToInterviewerDtoNoList.java
@@ -1,6 +1,6 @@
-package com.github.TheDreigon.JavaInterviewCalendarAPI.dto.converter;
+package com.github.TheDreigon.JavaInterviewCalendarAPI.converter.interviewer;
 
-import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.InterviewerDtoNoList;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.interviewer.InterviewerDtoNoList;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.Interviewer;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
