+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/dto/candidate/CandidateDto.java
@@ -1,5 +1,6 @@
-package com.github.TheDreigon.JavaInterviewCalendarAPI.dto;
+package com.github.TheDreigon.JavaInterviewCalendarAPI.dto.candidate;
 
+import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.AvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.Candidate;
 import jakarta.validation.constraints.*;
 import lombok.Data;
