+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/dto/candidate/CandidateDtoNoAvailability.java
@@ -1,16 +1,14 @@
-package com.github.TheDreigon.JavaInterviewCalendarAPI.dto;
+package com.github.TheDreigon.JavaInterviewCalendarAPI.dto.candidate;
 
-import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.Interviewer;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.Candidate;
 import jakarta.validation.constraints.*;
 import lombok.Data;
 
-import java.util.List;
-
 /**
- * The {@link Interviewer} data transfer object
+ * The {@link Candidate} data transfer object but with no availability list
  */
 @Data
-public class InterviewerDto {
+public class CandidateDtoNoAvailability {
 
     @Null
     private Integer id;
@@ -26,7 +24,4 @@ public class InterviewerDto {
     @Size(min = 3, max = 128)
     @Pattern(regexp = "[a-zA-Z]+", message = "Description should only use letters")
     private String description;
-
-    @Null
-    private List<InterviewerAvailabilityDto> interviewerAvailabilityDtoList;
 }
